azazel: Synchronize with vpsfree templates
Not copying the configuration to /etc since it worked without it in the past so I suspect this is for the benefit of nixos-container, which I do not plan to use.

Also had to change iproute to iproute2 since it was renamed <https://github.com/NixOS/nixpkgs/commit/100061531363d2b780baaedf07588adfad2c2148> and the old name was moved to aliases <https://github.com/NixOS/nixpkgs/commit/9378fdf87e0626e8c63a90a378c38444ff54808b>. But I cannot upstream this change now since it is not in NixOS stable.