azazel: Fix permissions for cpforum cache directory
The service starts under “root” user and the phpfpm daemon then lowers the euid to “cpforum”. But because systemd is not aware of that, the cache directory it creates does not have correct ownership.

Previously, we tried to use tmpfiles.d to set the correct permissions but that only worked if the service started before systemd-tmpfiles ran since each restart of the phpfpm-cpforum.service reset the permissions.

To fix this permanently, let’s correct the permissions after the service started. That way, systemd should be already done.

Fixes: https://github.com/jtojnar/nixfiles/issues/1

Weirdly, chmod needs to run before chown, otherwise it seems to change ownership to root:root. That makes no sense 🤷‍♀️