brian: Install dwarffs
This is complicated by the fact that brian is an Ubuntu box so we cannot use the NixOS service.

We are creating a systemd mount unit using home-manager but mount requires for `mount.fuse.dwarffs`
to exist on the default PATH to be able to use it: <https://github.com/libfuse/libfuse/issues/19>

Since home-manager cannot write to /sbin, we create a symlink in a stable location
and create a symlink from /sbin there manually using
`sudo ln -s $XDG_CACHE_HOME/mount.fuse.dwarffs /sbin/mount.fuse.dwarffs`.

Also install gdb from Nixpkgs since it is patched with dwarffs support.