azazel: Fix preklady.fan-club-penguin.cz/library/
I added a script tag loading Ruffle library to the page to render the Club Penguin library’s SWF files. But it turned out the JavaScript integration does not work (not sure if it ever worked with Flash), and clicking books navigates directly to their SWF files. This might have worked previously, when people still had Flash installed but we cannot inject the script to a standalone SWF file that browser will just try to download as a binary blob.

I therefore converted the `index.html` into a PHP page with a `/library/view/:file` endpoint that will send a HTML file with the flash file from URI embedded, and added `<base>` tag to make Ruffle’s `WebNavigatorBackend` open the new fake URLs.