flarum: Fix update script with --commit
When running `update` script with `--commit` flag, the update script
will run in a different worktree. Consequently, `repoSelf` will point
to a store path for the new worktree. On the other hand, `$0` comes
from `update.nix` so it should contain a path to `./update.sh`
in the original repository.

Now, this discrepancy would not be a problem – `update.nix` replaces
all paths to the original repository in `passthru.updateScript`
with their corresponding worktree paths.

But that scheme fails here because `update` takes the overlay list
from the `nixfiles`’s flake, so `./update.sh` is instead relative
to a store path corresponding to the repo.

As a result, relativization would produce a non-existent path.

Let’s use Nix to find out the path to `update.sh`
so that the store path is consistent.

Let’s also drop useless `--recursive` argument of `cp`.