theo: Disable “Unified Extensions Button” introduced in Firefox 109
It lists all enabled extensions, even the passive ones whose buttons are nigh useless.
And since the menu items cannot be hidden, the menu becomes very cluttered, with the amount of extensions I have installed.
The items cannot even be moved to the overflow menu so I am forced to either place the extensions I use infrequently on the toolbar, cluttering it, or to try to find them in the mess of the “unified” menu every time I do need them.
And even if they could be moved to “overflow”, I would still have the extra button for holding extensions whose action buttons I never press because that cannot be hidden either.
I could remove the overflow menu since I no longer have non-extension UI elements there (adding OpenSearch modules from address bar has been possible for a while and the Firefox Account menu is also accessible from the hamburger menu) so at least there would not be a second half-useless button on the toolbar but a single one is annoying enough.
Let’s disable this ugly Chromism for now.

https://www.askvg.com/disable-or-remove-extensions-button-from-firefox-toolbar/