Move session environment variables to a profile
Since we were previously setting the environment variables in a shell config,
they were not available to programs started by the DE such as Sublime Text.
As a result `CARGO_HOME` would be inconsistent between the terminal and the text editor,
causing the Cargo build directory to be invalidated every time one switched between the programs:

     Dirty …: the path to the source changed

Moving the GNUPGHOME to theo’s config since it requires socket hack that is only defined there.
And not like I am using gnupg right now anyway.