theo: Use disko for partitioning
I have to move the installation once in a while and partitioning always
makes me feel anxious because it involves performing a rather long
sequence of steps that include decisions I never remember.

Let’s encode the knowledge in Disko, which will hopefully make the next
move less stressful.

The configuration is based on
https://github.com/nix-community/disko/blob/72818e54ec29427f8d9f9cfa6fc859d01ca6dc66/example/luks-lvm.nix