azazel: Change php-fpm socket group ownership to `nginx`
`root` group in combination with `nginx` user is weird since the only user that inhabits the group `root` itself, and that bypasses ACL anyway.
Maybe it was meant to be used as a group that has no users as an extra safety in case the `mode` changes but hard to tell now.

Also, it only works because the php-fpm master process still runs as `root`, which will hopefully not be the case forever.

Either way, we want to allow other processes like the php-fpm Prometheus exporter to access it and, since ACL is not currently enabled in Nixpkgs, we need to use a group.

https://www.php.net/manual/en/install.fpm.configuration.php#listen-owner