common/profiles/core: Use file-roller for all archives
Nautilus is annoying, automatically extracting the archive in place.
That is slow, especially when most of the time, I just want to look at a single file.

Unfortunately, we cannot just replace it with `environment.systemPackages` because `gnome-session` adds its own `share` directory to `XDG_DATA_DIRS` so it would take precedence:

https://specifications.freedesktop.org/mime-apps/latest/file.html